requirements.txt
psycopg2-binary
python-dotenv
google-generativeai
pandas
tabulate
fastapi
uvicorn[standard]
streamlit
requests
pydantic
